+++ assets/pages/contributions-thankyou/components/marketingConsentContainer.jsx
@@ -1,0 +1,43 @@
+// @flow
+
+// ----- Imports ----- //
+
+import { connect } from 'react-redux';
+
+import { setGnmMarketing } from 'helpers/user/userActions';
+import { sendMarketingPreferencesToIdentity } from 'containerisableComponents/marketingConsent/helpers';
+import MarketingConsent from 'containerisableComponents/marketingConsent/marketingConsent';
+
+import type { Csrf as CsrfState } from 'helpers/csrf/csrfReducer';
+
+
+// ----- Component ----- //
+
+
+function mapDispatchToProps(dispatch) {
+  return {
+    onClick: (marketingPreferencesOptIn: boolean, email: string, csfr: CsrfState) => {
+      sendMarketingPreferencesToIdentity(
+        marketingPreferencesOptIn,
+        email,
+        dispatch,
+        csfr,
+        'CONTRIBUTIONS_THANK_YOU',
+      );
+    },
+    marketingPreferenceUpdate: (preference: boolean) => {
+      dispatch(setGnmMarketing(preference));
+    },
+  };
+}
+
+function mapStateToProps(state) {
+  return {
+    email: state.page.user.email,
+    marketingPreferencesOptIn: state.page.user.gnmMarketing,
+    consentApiError: state.page.marketingConsent.error,
+    csrf: state.page.csrf,
+  };
+}
+
+export default connect(mapStateToProps, mapDispatchToProps)(MarketingConsent);
